Spec256 should fetch instructions from CPU plane and sync every instruction
It turns out that the code in the graphics planes (GFX files) distributed
with Spec256 is corrupted and so the games rely on the SIMD nature of
Spec256 emulator to run correctly. We now fetch instructions from the CPU
memory, which is known to be good, and we change the granularity of CPU
state sync to one instruction. This makes ~6 out of 12 Spec256 games
work.