ld (HL), NN has two timing variants
There are two encodings of this instruction, one without prefix, the other
with ED prefix. The execution of the former takes 4 less T states, so
need to distinguish between the two. This fixes the border effect in
Darkstar highscore table being skewed. (Still, it remains offset from
the correct position.)