Don't interrupt twice
Since the reorg of video generator code we had most of the time
two interrupts slightly offset from each other, one generated by
the video generator code, the other by the main loop.
Leave it up to the video generator code to generate the Z80 interrupt.
In the future we should probably just pass the interrupt up
and the main module should relay it to the CPU/GPU and possibly
hook its bookkeeping off that as well.